+++ src/components/layout/list/Label.stories.tsx
@@ -1,0 +1,45 @@
+import React from "react";
+
+import type { StoryObj, Meta } from "@storybook/react";
+
+import { Label } from "./Label.js";
+import { List } from "./List.js";
+import { Item } from "./Item.js";
+
+const meta: Meta = {
+    title: "Components/Layout/List/Label",
+    component: Label,
+    tags: ["autodocs", "layout"],
+    argTypes: {
+        sub: {
+            type: "string",
+        },
+    },
+};
+
+type Story = StoryObj<typeof Label>;
+
+const Primary: Story = {
+    args: {
+        children: "I am a label",
+        sub: "Sub text",
+    },
+    render: (args) => {
+        return (
+            <List>
+                <Item>
+                    <Label {...args} />
+                </Item>
+                <Item>
+                    I am classic list item
+                </Item>
+            </List>
+        );
+    },
+};
+
+export {
+    Primary,
+};
+
+export default meta;