+++ CHANGELOG.md
@@ -3,10 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.20.1+]
+## [0.21.0] - 2022-06-13
 ### Breaking
 - `Section` no longer accepts `variant`, it has direct `horizontal` and `vertical` props now
+- `Item` from `List` no longer exported directly, use a `Item` property on `List` constructor
+- `List` accepts `inset` directly instead of `variant` prop
 ### Dev
+- `List` converted to stitches
 - `Section` moved to stitches
 - restored old dev mode